clearing service
